Pass a struct jpeg_upsampler_args to jsimd_*_upsample instead of the whole cinfo
The first argument to upsampling functions was

  j_decompress_ptr cinfo

Now it's replaced with two arguments

  JDIMENSION output_width,
  struct jpeg_upsampler_args args,

Similar to the last commit, output_width is outside the args structure
because it is a property of the whole decompression process.

The new args struct is this:

  struct jpeg_upsampler_args {
    int num_components;
    int max_v_samp_factor;
    J_COLOR_SPACE out_color_space;
    JSAMPLE *sample_range_limit;
    JDIMENSION output_scanline;

    /* FIXME: why does only mips/jsimd.c require this? */
    struct jpeg_upsampler *upsample;
  };

Note the FIXME; we'll fix it in another commit.